Spongebob now can interact with barrel
Still need to fix glitches.
After jumping onto barrel, doesn't fall after coming off.
The barrel picture is a square and the corners of it are transparent, so
when spongebob jumps on the corner, it looks like he is floating. Need
to fix that.
Need to add other objects and a life system